requirements.txt
absl-py==1.2.0
affine==2.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Cython==0.29.31
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.9.1
FALCONN==1.3.1
fastjsonschema==2.16.1
Fiona==1.8.21
fonttools==4.34.4
GDAL==3.4.3
geopandas==0.11.1
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.22.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipympl==0.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.8.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.8
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
patchify==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prefetch-generator==1.0.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyGObject==3.36.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyWavelets==1.4.1
pyzmq==23.2.0
qtconsole==5.4.0
QtPy==2.3.0
rasterio==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.2
six==1.14.0
sklearn==0.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.1
widgetsnbextension==3.6.1
zipp==3.8.1
